doc(DrawEvaluationBar): added documentation

+++ Graphics/evaluationBar.c
@@ -8,9 +8,13 @@ void DrawEvaluationBar(Rectangle *rectangle, const float *evaluation) {
     float clamped_score = atanf(*evaluation) / PI;
 
     DrawRectangleRec(*rectangle, BLACK);
-    DrawRectangle(rectangle->x, rectangle->y, rectangle->width, rectangle->height / 2.0f + rectangle->height / 2.0f * clamped_score, RAYWHITE);
+    DrawRectangle((int) rectangle->x,
+                  (int) rectangle->y,
+                  (int) rectangle->width,
+                  (int) (rectangle->height / 2.0f + rectangle->height / 2.0f * clamped_score),
+                  RAYWHITE);
 
-    DrawLine(rectangle->x, rectangle->y + rectangle->height / 2.0f,
+    DrawLine((int) rectangle->x, rectangle->y + rectangle->height / 2.0f,
              rectangle->x + rectangle->width, rectangle->y + rectangle->height / 2.0f,
              RAYWHITE);
 }
